# Conflicts:
#	src/main/java/com/fitbum/entidades/Notificacion.java
#	src/main/java/com/fitbum/entidades/estadisticas/Estadisticas.java
#	src/main/java/com/fitbum/repositorios/MenuRepositorio.java
#	src/main/java/com/fitbum/repositorios/usuarios/RolRepositorio.java
#	src/main/java/com/fitbum/repositorios/usuarios/UsuarioRepositorio.java

+++ src/main/java/com/fitbum/repositorios/NotificacionRepositorio.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-
-public interface NotificacionRepositorio extends JpaRepository<Notificacion,Integer> {
+public interface NotificacionRepositorio extends JpaRepository<Notificacion,Long> {
 }
